advanced/submissions/team-members/Patrick-Edosoma/requirements.txt
pydantic
python-dotenv
requests
fastapi
uvicorn
streamlit
gradio
pandas
tavily-python
pycountry
google-generativeai
